monday-course-manager/scripts/requirements.txt
requests==2.31.0
python-dotenv==1.0.1
monday==2.0.0
